Organize plant stuff, wip
